% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aws_dataVarObj_json.R
\name{get_aws_dataVarObj}
\alias{get_aws_dataVarObj}
\title{Create AWS variables JSON file for minute, hourly and daily data.}
\usage{
get_aws_dataVarObj(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Create AWS variables JSON file for minute, hourly and daily data.
}
