% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aws_parameters_json.R
\name{get_aws_parameters}
\alias{get_aws_parameters}
\title{Create AWS coordinates and parameters JSON file.}
\usage{
get_aws_parameters(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Create AWS coordinates and parameters JSON file.
}
