% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aws_variables_json.R
\name{get_aws_variables}
\alias{get_aws_variables}
\title{Create AWS variables JSON file.}
\usage{
get_aws_variables(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Create AWS variables JSON file.
}
