% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_aws_aggr_period.R
\name{initializeAWSAggrPeriod}
\alias{initializeAWSAggrPeriod}
\title{Initialize Table aws_aggr_period.}
\usage{
initializeAWSAggrPeriod(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.\cr}
}
\description{
Populate table \code{aws_aggr_period} from \code{adt_db} for hourly and daily data.
}
