% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_aws_dataDay.R
\name{update_dataDay_db}
\alias{update_dataDay_db}
\title{Update daily data database.}
\usage{
update_dataDay_db(aws_dir, dailyRainfall = FALSE, dailyRainObsHour = 8)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}

\item{dailyRainfall}{logical, if \code{TRUE} update daily rainfall only, if \code{FALSE} update other parameters.}

\item{dailyRainObsHour}{observation hour for daily rainfall data.}
}
\description{
Update daily data from AWS to ADT database.
}
