% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_aws_dataHour.R
\name{update_dataHour_db}
\alias{update_dataHour_db}
\title{Update hourly data database.}
\usage{
update_dataHour_db(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Update hourly data from AWS to ADT database.
}
