% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_aws_dataSP.R
\name{update_dataHour_sp}
\alias{update_dataHour_sp}
\title{Update hourly spatial data database.}
\usage{
update_dataHour_sp(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.\cr}
}
\description{
Update hourly spatial data from AWS to ADT database.
}
